+++ app/_styles/fonts.ts
@@ -1,0 +1,8 @@
+import { Montserrat } from 'next/font/google';
+
+export const montserrat = Montserrat({
+  subsets: ['latin'],
+  display: 'swap',
+  variable: '--font-montserrat',
+  weight: ['700'],
+});
