+++ app/_styles/globals.css
@@ -21,6 +21,8 @@
   }
   /* 부드러운 렌더링을 위한 속성 */
   body {
+    font-family: 'Pretendard Variable', sans-serif;
+    
     @apply min-h-screen bg-white text-gray-900;
     -webkit-font-smoothing: antialiased;
     -moz-osx-font-smoothing: grayscale;
@@ -32,3 +34,4 @@
     @apply ml-16 h-full min-h-screen bg-gray-FAFAFA tablet:ml-40 pc:ml-[18rem];
   }
 }
+
