--- app/fonts.ts
@@ -1,16 +1,0 @@
-import localFont from 'next/font/local';
-import { Montserrat } from 'next/font/google';
-
-export const pretendard = localFont({
-  src: '../public/fonts/PretendardVariable.woff2',
-  display: 'swap',
-  weight: '45 920',
-  variable: '--font-pretendard',
-});
-
-export const montserrat = Montserrat({
-  subsets: ['latin'],
-  display: 'swap',
-  variable: '--font-montserrat',
-  weight: ['700'],
-});
