+++ app/layout.tsx
@@ -1,5 +1,4 @@
 import '@/_styles/globals.css';
-import { pretendard } from './fonts';
 import { AuthProvider } from './context/AuthContext';
 
 export const metadata = {
@@ -20,9 +19,12 @@ export default function RootLayout({
           <meta charSet="UTF-8" />
           <meta name="viewport" content="width=device-width, initial-scale=1" />
           <title>{metadata.title}</title>
+          <link rel="preconnect" href="https://cdn.jsdelivr.net" crossOrigin="anonymous" />
+          <link rel="preload" as="style" crossOrigin="anonymous" href="https://cdn.jsdelivr.net/gh/orioncactus/pretendard@v1.3.8/dist/web/variable/pretendardvariable-dynamic-subset.css" />
+          <link rel="stylesheet" href="https://cdn.jsdelivr.net/gh/orioncactus/pretendard@v1.3.8/dist/web/variable/pretendardvariable-dynamic-subset.css" />
           <link rel="icon" href="/images/favicon.ico" />
         </head>
-        <body className={`${pretendard.variable}`}>{children}</body>
+        <body>{children}</body>
       </html>
     </AuthProvider>
   );
